+++ pkg-py/src/plotomics/__init__.py
@@ -50,4 +50,4 @@ __all__ = [
     "Embedding",
 ]
 
-__version__ = "0.1.0"
+__version__ = "0.2.0"